Revert "Небольшие исправления недочётов"
This reverts commit 9ca0437f8aee2f7567161e4c367f6b8ad0c4f157.

+++ CPSettings.cpp
@@ -50,7 +50,7 @@ void ControllerPlacementSettings::loadSettings(QSettings* const settings)
 	algorithm = settings->value("algorithm", def.algorithm).toInt();
 	useGraphviz = settings->value("useGraphviz", def.useGraphviz).toBool();
 
-	Lmax = settings->value("Lmax", def.Lmax).toDouble();
+	Lmax = settings->value("Lmax", def.Lmax).toInt();
 	CPerfomance = settings->value("CPerfomance", def.CPerfomance).toInt();
 	SPerfomance = settings->value("SPerfomance", def.SPerfomance).toInt();
 	CCost = settings->value("CCost", def.CCost).toInt();
